+++ src/utils/playwright.js
@@ -1,0 +1,5 @@
+module.exports = {
+  getDocumentScrollWidth: (page) =>
+    // eslint-disable-next-line no-undef
+    page.evaluate(() => window.document.documentElement.scrollWidth, {}),
+};